requirements.txt
blinker==1.8.2
boltons
brotlipy==0.7.0
certifi
cffi
charset-normalizer
click==8.1.7
cryptography
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
idna
itsdangerous==2.2.0
gunicorn==20.1.0
Jinja2==3.1.4
jsonpatch
jsonpointer==2.1
MarkupSafe==2.1.5
numpy
openpyxl==3.1.5
packaging
pandas
pluggy
pyjwt
pyOpenSSL
PySocks
python-dateutil==2.9.0
pytz==2024.1
requests
ruamel.yaml
six
toolz
tqdm
tzdata==2024.1
urllib3
Werkzeug==3.0.3
zstandard
Flask
gunicorn
Werkzeug
Authlib
google-cloud-storage
google-cloud-core
sdtp